restructures map stack to render via a screen component, removing all logic from file based navigation apart from the responsibility to render, passing state management and rendering of individual components to screen
